Implemented Alchemy session
